add world2chunk, world2chunk&block change ChunkPosition from uvec to ivec

+++ source/Render/World/Chunk/ChunkRenderer.cc
@@ -28,7 +28,7 @@ void ChunkRenderer::Render(const Camera &cam) {
   m_Program.Use();
   m_Program.SetUniform(0, mvp);
   for (auto &[pos, mesh] : m_Meshes) {
-    m_Program.SetUniform(1, mesh.position * static_cast<uint32_t>(CHUNK_SIZE));
+    m_Program.SetUniform(1, mesh.position * CHUNK_SIZE);
     mesh.Render();
   }
 }
